LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/type-change-in-keypress-listener-expected.txt



FAIL characters are not dropped when type changes to password during keypress assert_equals: expected "xyz" but got ""
